requirements.txt
Flask
Flask-SocketIO
tensorflow
pillow
googletrans==4.0.0-rc1
gTTS
